test(d): Introduce integration tests for waketimed
We have a "smoke" integration test - run the waketimed daemon, verify
3 threads got spawned, send sigterm, verify termination was
successful.

To spawn the D-Bus server, by default we connect to the session bus of
$DBUS_SESSION_BUS_ADDRESS of the user running the test. This required
redesigning how the toolbox container works. For D-Bus it is important
that the UID of the user connecting to session bus is the same as the
UID of the owner of the bus. So we can't keep using in-container
"rootless root" (UID 0) to connect to the session bus (e.g. UID 1000).
We have to start using `--userns=keep-id` parameter for the container.
This necessitated a bunch of dancing around in-container home dir and
mounts, but ultimately it seems to be working.