fix(d): Fix running waketimed in minimal config and test it
Waketimed engine couldn't enter Running state when no variables were
defined. This is now fixed.

Also it is now not considered a problem if one or both of the var_def
dirs do not exist. However, it is still considered a problem if any
can't be reached due to other reasons, e.g. permission errors.