change(d): Use seat instead of session for busy status
When a user would ssh into the system, then disconnect but e.g. leave
a lingering tmux session behind, from the perspective of logind that
login session would still be active, and in effect `wtd_user_busy`
would stay `true`.

We now look at logind seat, rather than session, for determining user
busy status. This means only the local login can influence whether
wtd_user_busy is true or false.

The downside of this approach is that even active (still connected)
ssh sessions will not be reflected in wtd_user_busy. Ideally we'd set
busy status to true due to active ssh connections, but not due to
lingering processes after disconnect. We may want to implement this in
the future.