change(d): Get rid of is_active var checks, allow polls to return None
It is not clear that is_active variable checks would be able to give a
correct answer at the time waketimed is starting, and then how many
times or until when should we repeat them etc.

So let's just get rid of the is_active concept for now, and poll every
variable each time, but make the poll function return an Option. If
the variable value cannot be determined, return None.