refact(d): Move PollVarFns and VarCreationContext to worker thread
Instead of sending the PollVarFns closures between threads, we can
make the worker thread own PollVarFns. This simplifies the program
structure.

It should also allow further simplification of PollVarFns - we should
get rid of the concept that poll_fn() returns closure that returns var
value, there should now be simply be a poll() function that will
return the var value.