feat: Configure and check allowed chassis types
There is now a config option 'allowed_chassis_types', a list of
chassis types (queryable via `hostnamectl chassis`) on which the
daemon will operate normally.

If the device's chassis type is not in the allowlist, the waketimed
engine enters Disabled state, in which no operation is done. (The
engine isn't even fully initialized, so no variable polling happens,
for example.) The daemon can only waits for termination signal in this
state.

This can be useful should some distros not have chassis-specific
variants and want to ship with waketimed installed and enabled. In the
disabled state, waketimed consumes very little memory and CPU.

The list of allowed types defaults to:

["convertible", "embedded", "handset", "tablet", "watch"]