refact: Consolidate into a single package
The original idea was that waketimed would provide a DBus API, and
there would be waketimectl to query that API, and a package
waketimed_core would be used for structure definitions common between
waketimed and waketimectl. However, it is not obvious if we'll ever
decide to use the API+ctl pattern for waketimed, so let's distribute
everything as a single Rust package for now. Should we decide to split
again, it shouldn't be too hard to extract the `core` module into its
own package.