dev: Set CARGO_HOME in tbx so that rust-analyzer can point to deps
Previously rust-analyzer worked with navigating to files around the
project. If we explicitly set CARGO_HOME to the path
$PROJECT_DIR/tmp/cargo (which we already use, we were just mounting it
in $HOME/.cargo), it will cause the paths to dependency source files
to be the same inside and outside of the tbx container. This means
that programmer can use rust-analyzer editor integration to open
implementation of functions in dependencies, not just in waketimed
itself.